# mate-platform-backend/contracts/openapi/generated/bundled.yaml
openapi: 3.1.0
info:
  title: Mate Platform API
  version: 1.0.0
  description: Canonical contract-first API for all Mate Platform domains.
  contact:
    name: Mate Platform API Governance
    email: api-governance@metaplatform.local
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  x-mate-owner: architecture-api-governance
servers:
  - url: /
    description: Gateway-relative production endpoint
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/msg/publish:
    post:
      summary: 发布消息
      tags:
        - msg
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResponse"
        "400":
          $ref: "#/components/responses/Error400"
        "500":
          description: 发布失败
      operationId: msgPostMsgPublish
      description: 发布消息
      x-mate-owner: platform-messaging
      x-mate-permission: msg.post
      x-mate-requirements:
        - FR-MSG-MSGPOSTMSGPUBLISH
      x-mate-implementation-status: implemented
  /api/v1/msg/topics:
    get:
      summary: 列出主题
      tags:
        - msg
      responses:
        "200":
          description: OK
        "400":
          $ref: "#/components/responses/Error400"
      operationId: msgGetMsgTopics
      description: 列出主题
      x-mate-owner: platform-messaging
      x-mate-permission: msg.get
      x-mate-requirements:
        - FR-MSG-MSGGETMSGTOPICS
      x-mate-implementation-status: placeholder
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    schemas-ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          pattern: ^E[0-9]{3}_[A-Z0-9_]+$
        message:
          type: string
          minLength: 1
        requestId:
          type: string
          minLength: 1
        details:
          type: object
          additionalProperties: true
...
    PublishRequest:
      type: object
      required:
        - topic
        - payload
      properties:
        topic:
          type: string
        payload:
          type: object
        partition_key:
          type:
            - string
            - "null"
        idempotency_key:
          type:
            - string
            - "null"
    PublishResponse:
      type: object
      properties:
        topic:
          type: string
        partition:
          type: integer
        offset:
          type: integer
        idempotency_hit:
          type: boolean
...
  responses:
...
    Error400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/schemas-ErrorResponse"